+++ Block/Adminhtml/Button/Generate.php
@@ -9,7 +9,7 @@ use Magento\Framework\Data\Form\Element\AbstractElement;
 /**
  * Class Generate
  *
- * @version 1.0.10
+ * @version 1.0.11
  * @author Improntus <http://www.improntus.com> - Ecommerce done right
  * @copyright Copyright (c) 2020 Improntus
  * @package Improntus\RetailRocket\Block\Adminhtml\Button
@@ -23,21 +23,21 @@ class Generate extends Field
     public function _getElementHtml(AbstractElement $element)
     {
         $element = null;
-        
+
         /** @var Button $buttonBlock  */
         $buttonBlock = $this->getForm()->getLayout()
             ->createBlock('Magento\Backend\Block\Widget\Button');
-       
+
         $url = $this->getUrl("retailrocket/generator/generate");
-            
+
         $data = [
             'class'   => 'improntus-retailrocket-generate-feed',
             'label'   => __('Generate Retail Rocket Feed Manually'),
             'onclick' => "setLocation('" . $url . "')",
         ];
-        
+
         $html = $buttonBlock->setData($data)->toHtml();
-        
+
         return $html;
     }
 }