+++ Block/Adminhtml/System/Config/Form/Date.php
@@ -8,7 +8,7 @@ use Magento\Framework\Stdlib\DateTime;
 /**
  * Class Date
  *
- * @version 1.0.10
+ * @version 1.0.11
  * @author Improntus <http://www.improntus.com> - Ecommerce done right
  * @copyright Copyright (c) 2020 Improntus
  * @package Improntus\RetailRocket\Block\Adminhtml\System\Config\Form